Improve Mutate methods
Mutate() is unexported. This will encourage suitable helper methods to be
written rather than testcases calling Mutate(), which is frowned upon.

Those occurrences which do not mutate the factory are named mutation(); those
few occurences which do mutate the factory (for good reason) are named
mutate().

If someone could kindly rewrite Kubernetes in Rust, immutability would be a
lot clearer.